requirements.txt
opencv-python<=4.6
grpcio
grpcio-tools
